Use fixed-width integers in Red Alert
Migrate the remaining `long` and `unsigned long` in the Red Alert sources:

- Radio messages carry their parameter (a TARGET) as int32_t through
  the whole Receive_Message chain and LParam.
- Version numbers, name and scenario CRCs, IPX register blocks,
  connection handlers and WOL IP addresses use int32_t/uint32_t.
- Prerequisite is uint64_t like the kStructFlag masks it is compared
  with; Get_Buildings builds its mask with a 64-bit shift. Ownable and
  Get_Owners use int32_t. Credit displays keep 64 bits to match
  HouseClass::Credits.
- The pathfinding cross product computes in int64_t.
- The generated WOL API header keeps its Win32 COM spellings under
  NOLINT, as do the two event-sink overrides that must match it.

Old-format INI triggers smuggled a new[]'d name string through CCPtr's
int ID, truncating the pointer on 64-bit builds. The name now waits in
TActionClass::PendingTriggerName until TriggerTypeClass::Read_INI
resolves it.

No serialized member changes width, so the save version is unchanged.

+++ src/ra/aircraft.cc
@@ -2691,7 +2691,7 @@ TARGET AircraftClass::New_LZ(TARGET oldlz) const {
  *=============================================================================================*/
 RadioMessageType AircraftClass::Receive_Message(RadioClass* from,
                                                 RadioMessageType message,
-                                                long& param) {
+                                                int32_t& param) {
   DCHECK_EQ(Aircraft.ID(this), ID);
   DCHECK(IsActive);
 
@@ -2800,7 +2800,7 @@ RadioMessageType AircraftClass::Receive_Message(RadioClass* from,
             if (cell == 0) {
               Transmit_Message(RADIO_OVER_OUT, from);
             } else {
-              param = static_cast<long>(::As_Target(cell));
+              param = static_cast<int32_t>(::As_Target(cell));
 
               /*
               **	Tell the potential passenger where it should go. If the
@@ -2809,7 +2809,7 @@ RadioMessageType AircraftClass::Receive_Message(RadioClass* from,
               */
               if (Transmit_Message(RADIO_MOVE_HERE, param, from) ==
                   RADIO_YEA_NOW_WHAT) {
-                param = static_cast<long>(As_Target());
+                param = static_cast<int32_t>(As_Target());
                 Transmit_Message(RADIO_TETHER);
                 if (Transmit_Message(RADIO_MOVE_HERE, param, from) !=
                     RADIO_ROGER) {
