+++ src/ra/foot.cc
@@ -1560,7 +1560,7 @@ bool FootClass::Restore_Mission() {
  *=============================================================================================*/
 RadioMessageType FootClass::Receive_Message(RadioClass* from,
                                             RadioMessageType message,
-                                            long& param) {
+                                            int32_t& param) {
   assert(IsActive);
 
   switch (message) {
@@ -1613,7 +1613,7 @@ RadioMessageType FootClass::Receive_Message(RadioClass* from,
     *motion, *	then it doesn't need further movement instructions.
     */
     case RADIO_NEED_TO_MOVE:
-      param = static_cast<long>(NavCom);
+      param = static_cast<int32_t>(NavCom);
       if (!Target_Legal(NavCom)) {
         return RADIO_ROGER;
       }
