+++ src/ra/interpal.cc
@@ -47,6 +47,7 @@
 #include "ra/interpal.h"
 
 #include <algorithm>
+#include <cstdint>
 
 #include "ra/ccfile.h"
 #include "ra/compat.h"
@@ -77,7 +78,8 @@ void Read_Interpolation_Palette(const char* palette_file_name) {
 
   if (palette_file.Is_Available()) {
     palette_file.Open(FileAccess::kRead);
-    palette_file.Read(&PaletteInterpolationTable[0][0], static_cast<long>(256) * 256);
+    palette_file.Read(&PaletteInterpolationTable[0][0],
+                      static_cast<int32_t>(256) * 256);
     palette_file.Close();
     InterpolationPaletteChanged = false;
   }
@@ -103,7 +105,8 @@ void Write_Interpolation_Palette(const char* palette_file_name) {
 
   if (!palette_file.Is_Available()) {
     palette_file.Open(FileAccess::kWrite);
-    palette_file.Write(&PaletteInterpolationTable[0][0], static_cast<long>(256) * 256);
+    palette_file.Write(&PaletteInterpolationTable[0][0],
+                       static_cast<int32_t>(256) * 256);
     palette_file.Close();
   }
 }
