+++ src/ra/ipxconn.h
@@ -94,7 +94,8 @@ class IPXConnClass : public ConnectionClass {
   static void Configure(uint16_t socket, int conn_num, ECBType* listen_ecb,
                         ECBType* send_ecb, IPXHeaderType* listen_header,
                         IPXHeaderType* send_header, char* listen_buf,
-                        char* send_buf, long handler_rm_ptr, int maxpacketlen);
+                        char* send_buf, int32_t handler_rm_ptr,
+                        int maxpacketlen);
 
   /*.....................................................................
   These routines tell IPX to start listening for packets, and to stop
@@ -186,7 +187,7 @@ class IPXConnClass : public ConnectionClass {
   the segment/offset to be computed by the caller gives the caller
   control over CS.)
   .....................................................................*/
-  static long Handler;
+  static int32_t Handler;
 
   /*.....................................................................
   This status flag tells us if Configure() has been called or not.
